+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Backwards incompatible changes
     - Minimum required elixir version is now "~> 1.2"
     - Minimum required erlang version is now "~> 18.0"
+    - Query builder support for "IF NOT EXISTS" has been removed
 
 ## v0.13.0 (2016-09-11)